requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
TA-Lib>=0.4.28
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0
python-telegram-bot>=20.0
scipy>=1.10.0
scikit-learn>=1.2.0
joblib>=1.2.0
tqdm>=4.65.0
colorlog>=6.7.0
streamlit>=1.25.0
plotly>=5.15.0
yfinance>=0.2.28
pandas_ta>=0.3.14b0
